+++ src/main/com/cqfn/degitx/simulator/Simulator.kt
