+++ src/main/com/cqfn/degitx/simulator/State.kt
