+++ src/main/com/cqfn/degitx/simulator/entities/Hardware.kt
@@ -1,4 +1,4 @@
-package com.cqfn.degitx.simulator
+package com.cqfn.degitx.simulator.entities
 
 /**
  * Represents server HW layer
