+++ src/main/com/cqfn/degitx/simulator/entities/Storage.kt
@@ -1,4 +1,4 @@
-package com.cqfn.degitx.simulator
+package com.cqfn.degitx.simulator.entities
 
 /**
  * Emulates data holder
