+++ src/main/com/cqfn/degitx/simulator/entities/impl/SATAStorage.kt
@@ -1,4 +1,6 @@
-package com.cqfn.degitx.simulator
+package com.cqfn.degitx.simulator.entities.impl
+
+import com.cqfn.degitx.simulator.entities.Storage
 
 /**
  * Data from benchmark: r500/w150 MB/s
